Update NavBar Buttons
- add RUN, CALIBRATE, update CONNECT
- add optional title prop to navbar
- update navbar in app, no run tab yet
