+++ components/src/icons/icon-data.js
@@ -37,6 +37,8 @@ export const MOVE = 'move'
 export const PLUS = 'plus'
 export const COPY = 'copy'
 export const MORE = 'more'
+export const CALIBRATE = 'calibrate'
+export const RUN = 'run'
 
 // icon data
 const ICON_DATA_BY_NAME = {
@@ -125,8 +127,8 @@ const ICON_DATA_BY_NAME = {
     path: 'M8.571 7.839l4.242 4.59-4.242 4.59 1.312 1.41 5.545-6-5.545-6z'
   },
   [CONNECT]: {
-    viewBox: '0 0 40 40',
-    path: 'M23.46 0l-.76 4.92A15.16 15.16 0 1 1 4.93 19.85 15.17 15.17 0 0 1 17.3 4.92L16.54 0A20.15 20.15 0 0 0 20 40a20.15 20.15 0 0 0 3.46-40zm-9.38 16.53C16.85 12.35 20 8 20 8s3.15 4.37 5.92 8.55c2.39 3.61 3.09 8.06.14 11.24a8.19 8.19 0 0 1-6 2.61H20a8.19 8.19 0 0 1-6-2.61c-3-3.19-2.31-7.65.08-11.26z'
+    viewBox: '0 0 24 24',
+    path: 'M17.3 9.9C15.1 6.5 12.5 3 12.5 3S9.9 6.5 7.7 9.9c-1.9 2.9-2.5 6.5-.1 9 1.3 1.4 3.1 2.1 4.9 2.1 1.8 0 3.6-.7 4.9-2.1 2.4-2.6 1.8-6.1-.1-9zm-4.9 9.4h-.3c-1.2-.1-2.4-.6-3.3-1.6-.7-.9-1.1-1.9-1.1-2.9 2.4.1 4.4-1.6 6.1-3.1.4-.4 1-.7 1.6-.7h.2c1.3.4 1.9 2.6 1.8 3.7-.1 2.6-2.3 4.7-5 4.6z'
   },
   [FILE]: {
     viewBox: '0 0 24 24',
@@ -179,6 +181,14 @@ const ICON_DATA_BY_NAME = {
   [MORE]: {
     viewBox: '0 0 24 24',
     path: 'M5.5 10A2.5 2.5 0 1 1 3 12.5 2.5 2.5 0 0 1 5.5 10zm7 0a2.5 2.5 0 1 1-2.5 2.5 2.5 2.5 0 0 1 2.5-2.5zm7 0a2.5 2.5 0 1 1-2.5 2.5 2.5 2.5 0 0 1 2.5-2.5z'
+  },
+  [CALIBRATE]: {
+    viewBox: '0 0 24 24',
+    path: 'M4.43 18.84v-.82H.99V19h1.99L.94 21.62v.86h3.5v-.97H2.36l2.07-2.67zm6.8-12.85h1.12l-.02-1.65 1.51-2.82h-1.27L11.8 3.2l-.81-1.68H9.68l1.55 2.76v1.71zm1.35 1.39h-1.55v6.28L5.2 16.85l.74 1.35 5.83-3.23 5.91 3.23.74-1.35-5.84-3.19V7.38zm9.04 12.79l1.37-2.15h-1.31l-.7 1.25-.69-1.25h-1.34l1.37 2.19-1.43 2.27h.2l1.13-.05.75-1.27.74 1.32h1.36l-1.45-2.31z'
+  },
+  [RUN]: {
+    viewBox: '0 0 24 24',
+    path: 'M18.83 11.58V1.89a.6.6 0 0 0-.61-.62H5a.64.64 0 0 0-.62.64v17.34a.62.62 0 0 0 .62.64h5.45a6.06 6.06 0 1 0 8.41-8.31zm-9.36 5.09a6 6 0 0 0 .4 2.1H5.54V2.48h12.14V11a6 6 0 0 0-8.2 5.66zm4.61 2.81v-5.64l3.72 2.76z'
   }
 }
