+++ components/src/index.js
@@ -21,6 +21,7 @@ export * from './forms'
 export * from './icons'
 export * from './instrument-diagram'
 export * from './structure'
+export * from './nav'
 export * from './lists'
 export * from './modals'
