+++ components/src/nav/NavButton.js
@@ -6,11 +6,19 @@ import styles from './navbar.css'
 import {type IconName, Icon} from '../icons'
 
 type NavButtonProps= {
+  /** optional click event for nav button */
   onClick?: (event: SyntheticEvent<>) => void,
+  /** position a single button on the bottom of the page */
   isBottom?: boolean,
+  /** classes to apply */
   className?: string,
+  /** disabled attribute (setting disabled removes onClick) */
   disabled?: boolean,
+  /** highlights navbutton if true */
   isCurrent?: boolean,
+  /** optional title to display below the icon */
+  title?: string,
+  /** Icon name for button's icon */
   iconName: IconName
 }
 
@@ -23,7 +31,8 @@ export default function NavButton (props: NavButtonProps) {
       disabled={props.disabled}
       onClick={props.onClick}
     >
-      <Icon name={props.iconName} />
+      <Icon name={props.iconName} className={styles.icon} />
+      {props.title && (<span className={styles.title}>{props.title}</span>)}
     </button>
   )
 }
