+++ components/src/nav/VerticalNavBar.js
