+++ components/src/nav/index.js
@@ -1,0 +1,11 @@
+// @flow
+// navigational components
+import VerticalNavBar from './VerticalNavBar'
+import NavButton from './NavButton'
+import SidePanel from './SidePanel'
+
+export {
+  VerticalNavBar,
+  NavButton,
+  SidePanel
+}
