+++ components/src/structure/index.js
@@ -2,9 +2,6 @@
 // structure components
 import PageTabs from './PageTabs'
 import TitleBar from './TitleBar'
-import VerticalNavBar from './VerticalNavBar'
-import NavButton from './NavButton'
-import SidePanel from './SidePanel'
 import Card from './Card'
 import RefreshCard from './RefreshCard'
 import LabeledValue from './LabeledValue'
@@ -15,9 +12,6 @@ export type {PageTabProps} from './PageTabs'
 export {
   PageTabs,
   TitleBar,
-  VerticalNavBar,
-  NavButton,
-  SidePanel,
   Card,
   RefreshCard,
   LabeledValue
